Inital support in the evo2 plugin for listing calendars
osync_member_call_plugin can now be used with a return value

the conflict handler now works like this:
the conflict gets checked. If a conflict occurs the handler
is called. After that the conflict gets checked again. if it is solved
it will get processed

the callback handler now support user_data arguments