Some changes to the formats:
there are now more xml support functions that can be used.
osxml_compare for example that compares 2 xmlDoc*s. the last
argument to this function is a array with weighted xpath
expression which can be used to test for similarity of 2 xml
trees.

Made some fixes to osync_group_load when the group has no name.
Had to fix the test data afterwards.