API changes:
when you register a converter you now dont need to
specify any flags. Just use the CONVERTER_DECAP etc
flags to specofy what you need.

The converter function now has a new argument:
osync_bool *free_input which you can set to TRUE
if you want opensync can free the input passed to the
converter (if you dont use a reference of the data, or
reference something inside this data)

The format plugin now register their functions with the
OSyncEnv, not the conv_env. The conv env is still used to
do the real work. When you spawn a new conv_env you have to
pass the pointer to a OSyncEnv. The register converters
etc of this env will then be resolved and stored in the conv_env

Other changes:

Reworked the way the path of converters is calculated.
detectors are now normal converters that can not always be invoked.
if a detector for a converter is set, opensync will invoke this
detector before calling the converter to make sure that we can make
the conversion.

Added a lot of tracing calls

The mappings now reside in the engine