Apply changes of revision 584 to trunk. svn merge http://svn.opensync.org/branches/opensync-alternate-field-scores -r583:584
Original log message:
Make osync_group_set_objtype_enabled() try to load plugin if necessary. Added an interface for requesting for "sink info", so that other parts of code that need sink information may check if it available and load it on demand. Added notes on documentation about requiring to instance the plugin before calling some functions, such as osync_member_set_objtype_enabled(), and marking changes to the interface of these functions to be able to return error as TODO