Send UID of the change back to engine on commit_change() reply
The plugin may have generated a new UID after committing the change. The
commit change reply will return the new UID of the change. This will solve
problems of the engine crashing after a delete: it was being caused by
an invalid mapping entry being set when a new UID was generated by the
plugin after an Add.