Pass group config dir to osplugin
The group we are synchronizing may have been loaded manually using
osync_group_load(). So change osplugin to get the group path from the
command line arguments, not the group name.

As a bonus, we make osplugin more efficient by avoiding loading
configuration for all groups because we already know where the group we
want is located.