Fix osplugin->enginepipe race problem
The IPC implementation uses only one pipe for incoming engine messages
(called "enginepipe"), but there is no mechanism to avoid two processes
from writing a message at the same time.

This fix this problem, by:

- Replacing engine->commandQueue and engine->incoming by
  engine->commands_to_self and engine->commands_from_self,
  for clearity
- Replacing client->incoming by client->commands_to_osplugin
- Creating client->commands_from_osplugin, to receive commands
  from osplugin

With this, we have a different command queue for each plugin, and we
won't have the race condition anymore.