Don't call osync_member_finalize() if osync_member_initialize() has failed
The plugin finalize() method isn't supposed to be called if initialize()
has failed.

Althought the engine needs to be able to send a FINALIZE command to
osplugin to finish the osplugin process. So, if the engine send us
a FINALIZE when osync_member_initialize() has failed, just exit from
osplugin.

This fixes osplugin crashes on the Plugin unit test, and maybe other
unit tests, also.