Set FD_CLOEXEC flag on lock file descriptor
We don't want the lock file to be kept open if the sync engine is finished
and other child processes (osplugin or processes started by libraries
used by plugins) keep running.

It solve some "Group is locked" problems that may happen if the
synchronization is aborted but osplugin isn't killed. But at a sync
engine abort, osplugin processes may still continue running (and may
cause problems on further synchronizations). This will be fixed in the
next commit.