Wait old osplugin process to terminate, when killing it
As we quickly start another osplugin process, the old osplugin process
may be still terminating after we sent a SIGTERM to it.

Fixing this by checking osync_queue_is_alive() after sending a SIGTERM
to osplugin. Wait at most 10 seconds for it to terminate, in this case
use SIGKILL.

Addding FIXME marks because the wait times are hardcoded in the code.