The Gnokii plugin should now correctly return the date/time in events entered on the phone. This may help if you experienced problems in the past that the time was a few hours off when syncing with the Gnokii plugin.
A crash was fixed which sometimes occured when syncing with Sunbird (and possibly other plugins)

The default type for items stored on the phone is now "meeting" (rather than "reminder" or "todo"). This ensures that start and end times are correctly set on the phone.

The timeouts (duration before OpenSync cancels the sync) has been increased. This helps if you transfer big calendars over Gnokii