Use osync_queue_new_pipes() for commands_{to,from}_self
As pipes aren't opened with O_NONBLOCK, the engine can't open both sides
of the .opensync/engines/enginepipe pipe because it will keep waiting
for the other side to be opened.

But as the 'enginepipe' pipe is used only for the engine
sending messages to itself, osync_queue_new_pipes() fits
perfectly for this task.

This also makes the code simpler as it doesn't need to touch the
filesystem. At least for the 'enginepipe' queues -- the osplugin queues
still live in the filesystem.

Signed-off-by: Eduardo Habkost <ehabkost@raisama.net>