Do proper message marshalling/demarshalling on get_change_data()
Use osync_{marshal,demarshal}_changedata() on get_change_data(), instead
of getting only the message data as a sequence of bytes. Synchronization
using file-sync plugin was broken because of this: file-sync data is a
struct containing pointers, it is not just a sequence of bytes.

Also, moving marshall/demarshall of change->has_data to
osync_marshal_changedata(), also, as it may be considered part of the
"change data", that is transfered on get_change_data().

Signed-off-by: Eduardo Habkost <ehabkost@raisama.net>