Move conversion from osync_member_commit_change() to osync_client_commit_change()
Do the conversion of the change data before sending it to osplugin,
on osync_client_commit_change().

This is just to make the osync_change_elevate() code work, because it
needs to use a method of the change format. However, moving the rest
of the more complex logic from osplugin to the sync engine is 
interesting, also.

Signed-off-by: Eduardo Habkost <ehabkost@silverbrick.(none)>