Set the startupfunc() event source as high priority
If the plugins register other event sources for the member event loop,
they may block the loop and make 'startupfunc' never be called. This
can make the engine hang while waiting for initialization of the
engine or a sync member. Changing the priority of the event source,
we make sure that startupfunc() is called as soon as possible.

This may fix the problem reported on ticket #241.