Free the engine queues on osengine_finalize(), not on osengine_free()
These queues are created on osengine_initialize(), so they should be
freed on osengine_finalize(), not on osengine_free().

This wasn't causing problems in the common case of a complete
synchronization process. But it was causing crashes for configuration
of incomplete groups (number of members < 2) on Kitchensync, because
osengine_initialize() was failing and the queues were not being created.