Fix a repeated bug in the wrappers that dgollub turned up:
Every time we return a pointer to an opensync object, we must increment 
its reference count, to prevent the object from being freed while 
python still holds a reference to it. This didn't happen for python 
sync plugins, but is quite possible for other python apps using the 
opensync API.

The reference should be decremented again when the python wrapper 
object is destroyed (as each destructor calls decref).