New implementation of configurable object formats.
Request "file" objformat from the engine. This makes storing of
configured objformats easier - since the engine takes care about
marshalling... and we can make use of the OSyncFile helpers.

For making 1:1 backups configure the <objformat> to "file" and disable
the converters in the engine...

e.g.: msynctool --sync BLUBB --disable-converters

Syncing with <objformat> set formatB and sourceformat formatA with
--disable-converters is non-sense ... because the converter gets involded.


Recommendation for UI Backup Wizard:

Create a spearate Backup group with the plugin for the device which
should get backuped and file-sync. file-sync with
<objformat>filer</objformat> (<objtype>data</objtype> (?)) and disable
the converters for each sync.

TODO: implement syncgroup config property to disable converters for the
group.