g_string_vprintf isn't able to deal with a string with NULL pointer on Solaris.
A string with null pointer is often passed to osync_trace function in which g_strdup_vprintf 
will deal with the string. On Solaris platform, g_strdup_vprintf isn't able to deal with 
a string with NULL pointer. On Linux platform, although it could be dealt with, I think it's not a good way.

In the patch, for the string argument passed to osync_trace function, its pointer will be checked 
at first if it is NULL or not. If it's null, the string "nil" instead of the pointer is passed.