Added plugin "mock-plugin-external" which is equal to "moto-sync". The only difference is the OSyncStartType for "mock-plugin-external"  is OSYNC_START_TYPE_EXTERNAL. ("moto-sync" has OSYNC_START_TYPE_THREAD (default)).
Make use of "mock-plugin-external" in the engine unit tests since nearly
every plugin got called with osync_client_run as external plugin process
- to test error condition in the plugin process and the engine handling
of failing plugin processes.

The "mock-plugin-external" is required since the OSyncPluginStartType is
currently not changeable since the StartType is hardcoded in the plugins
and to overwrite the OSyncPlugin value we have to touch OSyncPluginEnv
before osync_engine_initialized(). But this is not possible since the
OSyncPluginEnv got initialized in osync_engine_initialized() -> chicken
and the egg....